requirements.txt
   fastapi==0.115.8
   uvicorn==0.29.0
   gunicorn==20.1.0
   python-jose==3.4.0
   passlib>=1.7.4
   bcrypt>=3.2.0,<4.0.0
   python-multipart==0.0.20
   pydantic==2.10.6
   pydantic-settings==2.8.1
   sqlalchemy==2.0.38
   psycopg2-binary==2.9.9
   python-dotenv==1.0.1
   elasticsearch==7.17.9
   sentence-transformers==2.5.1
   torch>=2.0.0
   langchain==0.3.20
   langchain-community==0.3.19
   langchain-openai==0.3.8
   openai==1.61.1
   tiktoken>=0.7.0
   redis==4.3.4
   celery==5.4.0
   django-celery-beat==2.7.0
   beautifulsoup4>=4.12.3
   requests>=2.31.0,<3.0.0
   aiohttp>=3.9.3
   python-docx==1.1.2
   PyPDF2==3.0.1
   pytesseract==0.3.13
   pdf2image==1.17.0
   pillow==11.1.0
   PyMuPDF==1.23.26
   fastapi-mail==1.4.1
   aiofiles==23.2.1
   wikipedia==1.4.0
   arxiv==2.1.0
   fake-useragent==1.4.0
   playwright>=1.42.0
   trafilatura==2.0.0
   transliterate>=1.10.2
   python-magic
   mammoth
   rich
   langgraph
   langchain-core
   langchain-anthropic
   langchain-community
   langchain-openai
   pymysql>=1.1.0
   chardet>=5.0.0
   cryptography>=42.0.5
   pandas>=2.0.0
   tavily-python>=0.3.0
   google-generativeai>=0.5.0
